[MDL] First full integration of MDL with LLVM
This is the first public integration with the mainline LLVM Repo.  The
MDL language provides an alternate way (from TableGen) to describe a target
micro-architecture to the CodeGen and MC libraries. It provides all the
capability of Tablegen Schedules and Itineraries, but can efficiently
describe much more complex processors.

You can find the RFC with more detailed descriptions of this work at
[https://discourse.llvm.org/t/rfc-mdl-a-micro-architecture-description-language-for-llvm/66409] (from Nov 2022).
Detailed documentation about the design and the MDL language can be found in
the repo at llvm/docs/Mdl/.

This is the first integration of work that we plan to expand going
forward, and we'd like to land this to avoid bitrot and make the work
available to others. This integration is fully integrated - it can
extract microarchitecture information from TableGen, create equivalent
MDL descriptions, and compile and use that information in CodeGen and MC
libraries. This works for any target that has Schedules or Itineraries
(AArch64, AMDGPU, AMDGPU/R600, ARM, Hexagon, Lanai, Mips, PowerPC, RISCV,
Sparc, SystemZ, and X86).

MDL support is enabled using the LLVM_ENABLE_MDL cmake parameter. When enabled,
we build the MDL compiler and a tool to scrape necessary information from
tablegen files, then use this information in backend libraries instead
of using TableGen generated information. When not enabled (the default),
all the MDL-specific code is guarded by runtime flags that disable it.

About testing:
When you build with the MDL support disabled, compiler passes all check-all
tests.  When built with MDL support enabled - but disabled on the command line,
it also passes all check-all tests. When enabled on the command line
(the default), the compiler passes all but 175 tests (out of 90k+ tests).

We welcome comments and suggestions, and look forward to your feedback.

Differential Revision: https://reviews.llvm.org/D158790